+++ contacts/.eslintcache
@@ -1,1 +1,1 @@
-[{"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/index.js":"1","/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/App.js":"2"},{"size":208,"mtime":1606830599371,"results":"3","hashOfConfig":"4"},{"size":2813,"mtime":1606841827637,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"8"},"1texk13",{"filePath":"9","messages":"10","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/index.js",[],["11","12"],"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/App.js",[],{"ruleId":"13","replacedBy":"14"},{"ruleId":"15","replacedBy":"16"},"no-native-reassign",["17"],"no-negated-in-lhs",["18"],"no-global-assign","no-unsafe-negation"]
+[{"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/index.js":"1","/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/App.js":"2","/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/contactsList/contactList.js":"3","/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/ContactCard/ContactCard.js":"4"},{"size":208,"mtime":1606830599371,"results":"5","hashOfConfig":"6"},{"size":202,"mtime":1606844967939,"results":"7","hashOfConfig":"6"},{"size":2901,"mtime":1606845666997,"results":"8","hashOfConfig":"6"},{"size":460,"mtime":1606845648100,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1texk13",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/index.js",[],["19","20"],"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/App.js",[],"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/contactsList/contactList.js",[],"/mnt/c/Users/gonzalo/Documents/IRONHACK/WEB-DEVELOPMENT/Tareas/week7/week7-day2/lab-contacts/lab-react-ironcontacts/contacts/src/components/ContactCard/ContactCard.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
